[FIX] from float to double
[FIX] when the value is not found, return just the name

+++ src/Interpreter/Builtins/Arithmetic.hs
@@ -39,7 +39,7 @@ sub = arithmetic (-)
 multiplication :: Register -> [Tree] -> EvaluatedValue
 multiplication = arithmetic (*)
 
-arithmetic :: (Float -> Float -> Float) -> Register -> [Tree] -> EvaluatedValue
+arithmetic :: (Double -> Double -> Double) -> Register -> [Tree] -> EvaluatedValue
 arithmetic _    _   []            = ValueNumber 0
 arithmetic _    reg [value]       = checkIsNum $ evaluateValue (Context (reg, value))
 arithmetic func reg (value : xs)  = arithmetic' func (evaluateValue (Context (reg, value))) $ arithmetic func reg xs
@@ -48,7 +48,7 @@ checkIsNum :: EvaluatedValue -> EvaluatedValue
 checkIsNum value@(ValueNumber _)  = value
 checkIsNum _                      = throw ArgumentIsNotNumber
 
-arithmetic' :: (Float -> Float -> Float) -> EvaluatedValue -> EvaluatedValue -> EvaluatedValue
+arithmetic' :: (Double -> Double -> Double) -> EvaluatedValue -> EvaluatedValue -> EvaluatedValue
 arithmetic' func (ValueNumber a) (ValueNumber b)  = ValueNumber $ a `func` b
 arithmetic' _    _               _                = throw ArgumentIsNotNumber
 
@@ -60,10 +60,10 @@ divProcedure = arithmetic2 (/)
 modProcedure :: Register -> [Tree] -> EvaluatedValue
 modProcedure = arithmetic2 mod'
 
-arithmetic2 :: (Float -> Float -> Float) -> Register -> [Tree] -> EvaluatedValue
+arithmetic2 :: (Double -> Double -> Double) -> Register -> [Tree] -> EvaluatedValue
 arithmetic2 func reg [left, right] = arithmetic2' func (evaluateValue (Context (reg, left))) (evaluateValue (Context (reg, right)))
 arithmetic2 _    _   _             = throw InvalidNumberOfArguments
 
-arithmetic2' :: (Float -> Float -> Float) -> EvaluatedValue -> EvaluatedValue -> EvaluatedValue
+arithmetic2' :: (Double -> Double -> Double) -> EvaluatedValue -> EvaluatedValue -> EvaluatedValue
 arithmetic2' func (ValueNumber left) (ValueNumber right) = ValueNumber $ left `func` right
 arithmetic2' _    _                  _                   = throw ArgumentIsNotNumber
